python/requirements.txt
google-cloud
google-cloud-bigquery
google-cloud-storage